move resolve virtual entries
